same refactor
