+++ .npmignore
@@ -1,5 +1,3 @@
 /.*
 /*.md
-/*.html
-/app
-/vite.config.ts
+/apps
