+++ apps/docs/src/components/demo.ts
@@ -1,8 +1,8 @@
-import { commonStyle } from "@/app/styles.ts"
 import { fragment } from "@/lib/core.ts"
 import { css } from "@/lib/extra/css.ts"
 import { defineCustomTag } from "@/lib/extra/custom-tags.ts"
 import { html } from "@/lib/extra/html.ts"
+import { commonStyle } from "../styles.ts"
 
 const demoTag = defineCustomTag("x-demo")
 export function DemoWrapper() {
