+++ tsconfig.json
@@ -30,5 +30,5 @@
 			"@/*": ["./*"]
 		}
 	},
-	"include": ["app", "lib"]
+	"include": ["apps", "lib"]
 }
